feat: Add the tables Post and Comment and get-post route
The motive for this commit is to add the table post and comments to the
database and a route to access the post or comment.

This commit add the table Post and Comment the Post is a list of posts
and the Comment cam also be a comment or be a response to anothe comment
and the route get-post is to get a post by id and all the comments
related with that post.