feat: Add a route to list a Commit with it replies
the motive for this commit is to add a route that allow the server to
get a Commit by id and it replies as well to allow the user to easily
see what the communit are replaying to that commit.

This commit add a route to list a Commits by id with it replaies.