feat: Add a route to delete posts and commits
The motive for this commit is to add a route to allow the server to
delete posts and commits that are in the database.

This commit add two routes one to delete a post by id and another one to
delete a comment by id, both will delete the commits/replies connected
with they.