fix: Fix how the get-post sned the array of commits
As I was developing the Post's details I noticed that I woulden't be
able to list the array of commits like we see in the site TabNews in
order to fix it I decided to organize the commits in the server.

This commit fix the array of commits to order by the newst to the oldest
and fix it three.