feat: Add the function to create comments and replies.
The mothive for this commit is to put ina final version the
infrastructure to add commits and replies knowning that the last one was
a text.

This commit add the infrastructure to add commit/reply and also
refresh whem ther's a new comment/reply in that post.

PS: I'm looking forward to add a pub-sub method here.