feat: Add the cherish and depreciate buttons.
The potive for this commit is to allow the user to cherish or depreciate
a post or comment to allow the other to know if that one is adding or
not value to the post and also help to insentive the users to work on
good good posts and comments to add to the readers.

This commit add the functinos to cherish or depreciate to posts and
comments it buttons are in the CarretUp/Down in the left of the
post/comment.