feat: Add create user/beeKeeper with crypted data.
The motive for this commit is to add a create user/beeKeeper route with
encrypted personal info. It also changes the promCoins in the prisma to asfCash cuz
semantic. And fixed the ENCRYPTION_KEY that is not a url.

This commit add create user/beeKeper route. Changh env.ts and
encrypt.ts.