feat: Add userLogon with token and crytography.
The motive for this commit is to add the Logon route with token and
learn how to use token and cryptograth to protect personal data in the
database, as well as to protect the token in the front end. not mentioning to
put the .env to work.

This commit add a logon route with token, my first time using token!
And cryptograthy funtions to use in the server. It also add a
tsconfig because I suddenly started having problens with the requires.
I have updated the @types/node trying to fix the same problem. And the
.env is now namede .env.example. Yes, I forgot to rename it before.