feat: Add the user things, login and docker.
The motive for this commit is to add a login sisten and the things the user need to interact in the aplication, to fix the cherish/depreciate routes and to add a coins/cash sisten for incentive the user to create relevant contents. it also add a docoker file/compose to run the aplication in docker, keep in mind that the docker file is not otimized yed.

This commit fix the cherish/depreciate route, add a docker file/compose and add a coins/cash sisten in the aplication. It also explit the PostContext in two, one for posts and another for user, eaxe one have it on connection with the webSocket.