feat: Add a userBar in home
The motive for this commit is to add a userBar in home to facilite the user to identify the events that is set in their state to help the community to interact with each other.

This commit add a userBar in home, add a collum state in the database and change the asfCoins in the comment and post to value.